TWINS-732: support new storage for statisters

+++ core/src/main/java/org/twins/core/dao/twin/TwinFieldHeadSumCountProjection.java
@@ -1,6 +1,7 @@
 package org.twins.core.dao.twin;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
-public record TwinFieldHeadSumCountProjection(UUID headTwinId, Double sum, Long count) {
+public record TwinFieldHeadSumCountProjection(UUID headTwinId, BigDecimal sum, Long count) {
 }
